pyproject.toml
[project]
name = "document-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.70.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.2",
]
